# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PushUp Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserInputFitnessLevel } from "./userInputFitnessLevel";
import type { UserInputMainGoal } from "./userInputMainGoal";

export interface UserInput {
  /** @minLength 1 */
  name: string;
  fitnessLevel: UserInputFitnessLevel;
  /** @minimum 0 */
  currentMaxPushups: number;
  /**
   * @minimum 1
   * @maximum 7
   */
  weeklyAvailabilityDays?: number;
  mainGoal: UserInputMainGoal;
  reminderPreference?: string;
  wantsAiGoals?: boolean;
  wantsCompetition?: boolean;
}