# lib/api-zod/src/generated/types/userStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PushUp Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserStats {
  userId: number;
  totalPushups: number;
  todayReps: number;
  weeklyTotal: number;
  averageRepsPerSet: number;
  currentStreak: number;
  bestSet: number;
  progressScore: number;
  weeklyReps: number;
  streakBonus: number;
  personalBestBonus: number;
  challengeBonus: number;
  verifiedRepsTotal: number;
  unverifiedRepsTotal: number;
  weeklyVerifiedReps: number;
  weeklyUnverifiedReps: number;
}