# lib/api-zod/src/generated/types/workout.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PushUp Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Workout {
  id: number;
  userId: number;
  date: Date;
  totalReps: number;
  verifiedReps: number;
  unverifiedReps: number;
  /** @nullable */
  cameraAngle?: string | null;
  sets: number;
  averageReps: number;
  variation: string;
  usedCamera: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}