# lib/api-zod/src/generated/types/workoutInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PushUp Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutInputCameraAngle } from "./workoutInputCameraAngle";

export interface WorkoutInput {
  userId: number;
  /** @minimum 0 */
  totalReps: number;
  /** @minimum 0 */
  verifiedReps?: number;
  /** @minimum 0 */
  unverifiedReps?: number;
  cameraAngle?: WorkoutInputCameraAngle;
  /** @minimum 1 */
  sets: number;
  variation: string;
  usedCamera: boolean;
  notes?: string;
  date?: Date;
}